requirements.txt
feedparser
python-dateutil
pyyaml
requests
python-dotenv
openai
